Adding sensor stubs. Needs to be rethought a bit.
I am working on integrating normalized sensors into the social
structure to feed information to tribes. As tribes evolve the content
they are fed will adapt to the most relevant sensors in the tribes
network. The idea is to make all parts of the system measurable and
quantifiable. We need to be able to cut cross sections into the flow of
data and project future cross sections. The idea is to predict flow and
adapt to it before it happens.